G64 Problem jetzt auf 64Bit Systemen. Problem mit der Bitlänge von unsigned long auf verschidenen Systemen 32/64 Bit. Habe das jetzt so gelöst das ich dort als Datentyp uint32_t genommen habe. Da gibt es jetzt kein Missverständnis mehr!

+++ src/floppy1541_class.cpp
@@ -250,8 +250,8 @@ bool Floppy1541::LoadDiskImage(char* filename)
         char  version;
         char  trackanzahl;
         unsigned short  tracksize;
-        unsigned long trackpos[84];
-        unsigned long trackspeed[84];
+        uint32_t trackpos[84];
+        uint32_t trackspeed[84];
 
         // Diskwechsel simulieren
         StartDiskChange();
